rename -funcs to -fn
