+++ lib/-fn.js
