+++ lib/api/init.js
@@ -1,8 +1,8 @@
 'use strict';
 
 const Promise = require('bluebird');
-const $ = require('../-funcs');
 const co = Promise.coroutine;
+const $ = require('../-fn');
 const READY = '_ready';
 
 module.exports = co(function * () {
