+++ lib/reporter.js
@@ -1,7 +1,7 @@
 'use strict';
 
+const print = require('./-fn').formatTime;
 const fmt = require('./fmt');
-const fmtTime = require('./-funcs').formatTime;
 
 const evts = {
 	add: 'added',
@@ -56,7 +56,7 @@ module.exports = function () {
 		})
 
 		.on('task_complete', obj => {
-			const t = fmtTime(obj.time);
+			const t = print(obj.time);
 			$.log(`Finished ${fmt.complete(obj.task)} in ${fmt.time(t)}`);
 		})
